+++ CHANGELOG.rst
@@ -2,6 +2,59 @@
 Changelog for package s3000_laser
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.8.2 (2018-10-23)
+------------------
+* fixes for narrowing issues with c++11
+* Contributors: Tracey Spicer
+
+0.8.1 (2017-12-07)
+------------------
+* CORE-9158: Better interrupt byte check.
+* Contributors: Mike Purvis
+
+0.8.0 (2017-11-29)
+------------------
+* CORE-8925: Try harder to reset laser after detecting error.
+* Contributors: Mike Purvis
+
+0.7.2 (2017-05-03)
+------------------
+* CORE-7222: Recover gracefully from bad TX comms
+* CORE-6899: Re-poll code when H or G.
+* Contributors: Mike Purvis
+
+0.7.1 (2017-03-10)
+------------------
+* CORE-6843: Fix SEG_N value, now can properly recover N1 errors.
+* Add missing boost/scoped_ptr include.
+* Contributors: Mike Purvis
+
+0.7.0 (2017-03-01)
+------------------
+* CORE-6646: Error counters
+* CORE-6647, CORE-6645: Updates to fault resetting.
+* CORE-6505, CORE-6506: Re-enable error diagnostics, with error resets.
+* Contributors: Mike Purvis
+
+0.6.0 (2016-12-16)
+------------------
+* Better distinguish unset from zero.
+* Report SCID values as hex.
+* Bump version number.
+* Contributors: Mike Purvis
+
+0.1.7 (2016-11-14)
+------------------
+* Add missing s3000_laser to install targets.
+* Contributors: Justin Godson
+
+0.1.6 (2016-11-14)
+------------------
+* Add block check call on rostopic to s3000_laser.cc
+* Split ReadLaser and move read delay out of driver.
+* Add block diagnostics check functionality to s3000 driver.
+* Contributors: Justin Godson
+
 0.1.5 (2016-05-09)
 ------------------
 * Parameterize frequency warning tolerance, change default from 5% to 20%.